### 2023-12-28 #620 @mpru
Aprendiendo programación en R con la robot Karel
Nombre de la Persona Encargada: Marcos Prunello
Usuario GitHub de la Persona Encargada: @mpru
Repositorio: https://github.com/mpru/karel
Versión Enviada: 0.1.1.9000
Tipo de Envio: Estándar
Editora: @maurolepore
Revisores: @vjimenez9, @joelnitta

Archivo: TBD
Versión Aceptada: TBD
Idioma: es
---

-   Pega el archivo DESCRIPTION completo dentro del siguiente bloque de código.

```
Package: karel
Title: Learning programming with Karel the robot
Version: 0.1.1.9000
    person(given = "Marcos",
           family = "Prunello",
           email = "marcosprunello@gmail.com",
           comment = c(ORCID = "0000-0002-9611-527X"))
Description: This is the R implementation of Karel the robot, a programming
  language created by Dr. R. E. Pattis at Stanford University in 1981. Karel is
  an useful tool to teach introductory concepts about general programming, such
  as algorithmic decomposition, conditional statements, loops, etc., in an
  interactive and fun way, by writing programs to make Karel the robot achieve
  certain tasks in the world she lives in. Originally based on Pascal, Karel
  was implemented in many languages through these decades, including 'Java', 'C++',
  'Ruby' and 'Python'. This is the first package implementing Karel in R.
    gganimate,
    cli
Language:
  en,
  es
```

## Alcance

- Por favor, indica qué categoría(s) aplican a este paquete. Las puedes encontrar en nuestras [políticas de inclusión de paquetes (Inglés)](https://ropensci.github.io/dev_guide/policies.html#package-categories). Por favor, tilda todas las apropiadas. Si no estás seguro, te sugerimos que comiences un pre-envío.

	- [ ] recuperación de datos (data retrieval)
	- [ ] extracción de datos (data extraction)
	- [ ] munging de datos (data munging)
	- [ ] disposición o declaración de datos (data deposition)
    - [ ] validación y prueba de datos (data validation and testing)
	- [ ] automatización de flujos de trabajo (workflow automation)
	- [ ] control de versiones (version control)
	- [ ] manejo de citas y bibliometría (citation management and bibliometrics)
	- [ ] envoltorios de software científico (scientific software wrappers)
	- [ ] herramientas para trabajo de campo y reproducibilidad (field and lab reproducibility tools)
	- [ ] ligamientos con software de base de datos (database software bindings)
	- [ ] datos geoespaciales (geospatial data)
	- [ ] análisis de texto (text analysis)

- Explica cómo y por qué el paquete encaja dentro de estas categorías (1 a 3 oraciones):

	Este paquete no entra en ninguna categoría, ya que está relacionado con Educación. Fue aceptado para el Programa Campeones, bajo la supervisión de @yabellini.

- ¿Cuál es la audiencia esperada y las aplicaciones científicas de este paquete?

	Este paquete es utilizado por docentes para enseñar conceptos introductorios sobre programación, como descomposición algorítmica, declaraciones condicionales, bucles, etc., a estudiantes que posteriormente utilicen R como lenguaje de programación para otras tareas. Ha sido creado con un enfoque multilingüe que actualmente incluye los idiomas español e inglés, pero que puede ser expandido a otros de forma sencilla.

- ¿Hay otros paquetes de R que logren el mismo objetivo? Si los hay, ¿cómo se diferencian del tuyo, o alcanzan nuestro [criterio del mejor de su categoría (documento en Inglés)](https://ropensci.github.io/dev_guide/policies.html#overlap)?

	No actualmente, según mi conocimiento.

-   (Si aplica) ¿Tu paquete cumple con nuestras [guías de _Ética, Privacidad de Datos e Investigación de Sujetos Humanos_ (documento en Inglés)](https://devguide.ropensci.org/policies.html#ethics-data-privacy-and-human-subjects-research)?

	No aplica.

- Si ya has hecho una consulta de pre-envío, por favor pega el enlace al _issue_ correspondiente, una publicación del foro, u otra discusión. Alternativamente, etiqueta al editor (con @tag) con el que te contactaste.

	No realicé consulta de pre-envío.

-   (Si aplica) Explique las razones de los [elementos `pkgcheck`](https://docs.ropensci.org/pkgcheck/) que su paquete no puede pasar.

	pkgcheck pasa cuando lo ejecuto localmente. He observado [dos advertencias](https://github.com/mpru/karel/actions/runs/7347980615) en el CI, pero creo que están relacionadas con pkgcheck y GitHub Actions, y no relacionadas directamente con mi paquete. En cuanto al estilo, obtengo la sugerencia de evitar largas líneas de código. Sin embargo, las únicas líneas de código que superan los 80 caracteres de ancho corresponden a cadenas de texto con mensajes en diferentes idiomas para los usuarios, dentro de las cuales prefiero no incluir saltos de línea.

## Revisiones Técnicas

Tilda los siguientes items para confirmar que los has completado:

- [x] He leído la [guía sobre paquetes de rOpenSci](https://devguide.ropensci.org/building.html).
- [x] He leído la [guía para autores](https://devdevguide.netlify.app/authors-guide.html) y pienso mantener este paquete durante al menos 2 años o encontrar un reemplazo.

Este paquete:

- [x] No viola los Términos de Servicio de ningún servicio con los que interactúa.
- [x] Tiene una licencia aceptada por CRAN y OSI.
- [x] Contiene un [archivo README con instrucciones para instalar la versión de desarrollo](https://ropensci.github.io/dev_guide/building.html#readme).
- [x] Incluye [documentación con ejemplos para todas las funciones, creada con roxygen2](https://ropensci.github.io/dev_guide/building.html#documentation).
- [x] Contiene una viñeta (_vignette_) con ejemplos de sus funciones esenciales y su uso.
- [x] Tiene una [suite de tests (documento en Inglés)](https://devguide.ropensci.org/building.html#testing).
- [x] Tiene [integración contínua (documento en Inglés)](https://devguide.ropensci.org/ci.html), incluyendo reporte de cobertura de tests.

## Opciones de Publicación

- [x] ¿Tienes intenciones de subir este paquete a CRAN?
- [ ] ¿Tienes intenciones de enviar este paquete a Bioconductor?

- [ ] ¿Deseas enviar un Artículo de Aplicaciones sobre tu paquete a [Methods in Ecology and Evolution (documento en Inglés)](http://besjournals.onlinelibrary.wiley.com/hub/journal/10.1111/(ISSN)2041-210X/)? Si es así:

<summary>Opciones para MEE</summary>

- [ ] Este paquete es novedoso y será de interés para la mayoría de lectores de la revista.
- [ ] El manuscrito que describe el paquete no tiene más de 3000 palabras y está escrito en Inglés.
- [ ] Tienes intenciones de archivar el código del paquete en un repositorio a largo plazo, que cumple los requerimientos de la revista (mira las [Políticas de Publicación de MEE (documento en Inglés)](http://besjournals.onlinelibrary.wiley.com/hub/journal/10.1111/(ISSN)2041-210X/journal-resources/policy-on-publishing-code.html))
- (*Alcance: Considera los [Objetivos y Alcance de MEE (documento en Inglés)](http://besjournals.onlinelibrary.wiley.com/hub/journal/10.1111/(ISSN)2041-210X/aims-and-scope/read-full-aims-and-scope.html) para tu manuscrito. No otorgamos garatías de que tu manuscrito esté en el ámbito de MEE.*)
- (*Aunque no es requerido, recomendamos tener un manuscrito completamente preparado y en Inglés, al momento de enviar.*)
- (*Por favor, no envíes tu paquete de forma separada a Methods in Ecology and Evolution*)

## Código de Conducta

- [x] Estoy de acuerdo en cumplir el [Código de Conducta de rOpenSci (documento en Inglés)](https://ropensci.github.io/dev_guide/policies.html#code-of-conduct) durante el proceso de revisión, y en mantener mi paquete si éste es aceptado.
