### 2024-01-03 #620 @maurolepore
@mpru, es un placer editar tu paquete.

Pronto empezaré a trabajar en la [lista de verificación](https://devdevguide.netlify.app/es/editortemplate.es).

Mientras tanto:

 - [ ] ¿Confirmas que la revisión debe ser en castellano?
 - [ ] Por favor, sugiere 3 revisores/as. En rOpenSci, generalmente convocamos a no más de 1 de las personas sugeridas, pero utilizamos el resto para comprender mejor el perfil que crees que añadiría valor a tu paquete. Por favor evita [conflictos de interés](https://devguide.ropensci.org/policies.html#coi).
