### 2024-01-07 #620 @maurolepore
@mpru,

Seguro ayudo con eso. Nuestra guia ofrece varias ideas sobre donde puedo buscar revisore/as:

https://devdevguide.netlify.app/es/softwarereview_editor.es#where-to-look-for-reviewers

Yo usare esa misma guia para buscar revisore/as por mi cuenta, pero nos interesa tambien tu propia opinion.
