### 2024-01-07 #620 @maurolepore
Dear @mpru,thanks for your work. Editor checks were super smooth.

You'll see some comments that require your attention. They are labeled ml01, ml02, and so on. The one(s) starting with a checkbox are required. The one(s) starting with a bullet are just for your consideration.

---

FIT

rOpenSci accept this package as part of the Champions program.

VIGNETTES

* ml01. In the article "Getting started" consider removing "new" so that this sentence is still valid many years from now:

> karel is a **new** R package ...

TESTS

- [ ] ml02. Please explain if and how the package tests interactivity.

* ml03. I see few general tests. Consider spliting them in to more, smaller, specific tests so that it's easier
to detect the problem when a test fails.

A search for "test_that(" shows that the test titles seem too general -- https://github.com/search?q=repo%3Ampru%2Fkarel%20test_that(&type=code

And I see many expectations per tests, e.g.:

https://github.com/mpru/karel/blob/master/tests/testthat/test-actions.R

> You want to arrange things such that, when a test fails, you’ll know what’s wrong and where in your code to look for the problem. This motivates all our recommendations regarding file organisation, file naming, and the test description. Finally, try to avoid putting too many expectations in one test - it’s better to have more smaller tests than fewer larger tests.

-- https://r-pkgs.org/testing-basics.html#test-organisation

* ml04. I see the expected error is not specific.

> Usually you care about two things when testing an error:
>
> * Does the code fail? Specifically, does it fail for the right reason?
> * Does the accompanying message make sense to the human who needs to deal with the error?

-- https://r-pkgs.org/testing-basics.html#expectations

* ml05. "Your test files should not include these `library()` calls." -- https://r-pkgs.org/testing-basics.html#test-organisation

For example see  https://github.com/search?q=path%3Atests%2Ftestthat+repo%3Ampru%2Fkarel+library%28karel%29&type=code
