### 2024-01-07 ae9a499
changed sentence in getting started vignette, both es and en
