+++ tests/testthat/test-actions.R
@@ -1,5 +1,3 @@
-library(karel)
-
 test_that("Karel's actions perform as expected, using Spanish functions", {
   generar_mundo("mundo001")
   karel_env <- get_pkg_env()
