+++ tests/testthat/test-actions.R
@@ -1,79 +1,0 @@
-test_that("Karel's actions perform as expected, using Spanish functions", {
-  generar_mundo("mundo001")
-  karel_env <- get_pkg_env()
-  avanzar()
-  expect_equal(karel_env$x_now, 2)
-  expect_equal(karel_env$y_now, 1)
-  juntar_coso()
-  expect_equal(karel_env$beepers_any, 0)
-  girar_izquierda()
-  expect_equal(karel_env$dir_now, 2)
-  avanzar()
-  poner_coso()
-  expect_equal(karel_env$beepers_any, 1)
-})
-
-test_that("Karel's actions perform as expected, using English functions", {
-  generate_world("mundo001")
-  karel_env <- get_pkg_env()
-  move()
-  expect_equal(karel_env$x_now, 2)
-  expect_equal(karel_env$y_now, 1)
-  pick_beeper()
-  expect_equal(karel_env$beepers_any, 0)
-  turn_left()
-  expect_equal(karel_env$dir_now, 2)
-  move()
-  put_beeper()
-  expect_equal(karel_env$beepers_any, 1)
-})
-
-test_that("The program stops running if the user made a mistake", {
-  generate_world("mundo001")
-  turn_right()
-  expect_error(move())
-  expect_error(turn_right())
-  expect_error(turn_left())
-  expect_error(turn_around())
-  expect_error(move())
-  expect_error(put_beeper())
-  expect_error(pick_beeper())
-  expect_error(run_actions())
-})
-
-test_that("Karel can't move through walls", {
-  generate_world("mundo001")
-  turn_around()
-  expect_error(move())
-  generate_world("mundo001")
-  move()
-  turn_around()
-  move()
-  turn_around()
-  move()
-  move()
-  expect_error(move())
-  generate_world("mundo001")
-  turn_left()
-  move()
-  move()
-  turn_around()
-  move()
-  turn_around()
-  move()
-  move()
-  expect_error(move())
-})
-
-test_that("Karel handles correctly putting and picking beepers", {
-  generate_world("mundo001")
-  expect_error(pick_beeper())
-  world_test <- list(nx = 6, ny = 4,
-                     hor_walls = dplyr::tibble(x = 3, y = 1, lgth = 3),
-                     ver_walls = dplyr::tibble(x = 3, y = 0, lgth = 1),
-                     karel_x = 1, karel_y = 1, karel_dir = 1,
-                     beepers_x = 2, beepers_y = 1, beepers_n = 1,
-                     beepers_bag = 0)
-  generar_mundo(world_test)
-  expect_error(put_beeper())
-})
