### 2024-01-08 e102b49
opened test-conditions in 2 files, one for each language, and made unit tests more specific
