+++ pkgdown.yml
@@ -14,7 +14,7 @@ articles:
   d_controlstructures_es: d_controlstructures_es.html
   e_examples_en: e_examples_en.html
   e_examples_es: e_examples_es.html
-last_built: 2024-01-07T22:59Z
+last_built: 2024-01-08T01:26Z
 urls:
   reference: https://mpru.github.io/karel/reference
   article: https://mpru.github.io/karel/articles
