### 2024-02-12 bc40758
fixing bug found while improving test units
