+++ tests/testthat/test-set_up.R
@@ -1,5 +1,5 @@
-test_that("Checking provided nx and ny.", {
-  # Create one world
+test_that("a new world provided as a list by the user meets all the criteria.
+          Checking that Karel's initial direction is indicated.", {
   world_test <- list(nx = 6, ny = 4,
                      hor_walls = dplyr::tibble(x = 3, y = 1, lgth = 3),
                      ver_walls = dplyr::tibble(x = 3, y = 0, lgth = 1),
@@ -8,7 +8,10 @@ test_that("Checking provided nx and ny.", {
                      beepers_x = 2, beepers_y = 1, beepers_n = 1,
                      beepers_bag = Inf)
   expect_error(generar_mundo(world_test))
+})
 
+test_that("a new world provided as a list by the user meets all the criteria.
+          Checking that the number of avenues (nx) is correct.", {
   world_test <- list(nx = 6.4, ny = 4,
                      hor_walls = dplyr::tibble(x = 3, y = 1, lgth = 3),
                      ver_walls = dplyr::tibble(x = 3, y = 0, lgth = 1),
@@ -24,7 +27,10 @@ test_that("Checking provided nx and ny.", {
                      beepers_x = 2, beepers_y = 1, beepers_n = 1,
                      beepers_bag = Inf)
   expect_error(generar_mundo(world_test))
+})
 
+test_that("a new world provided as a list by the user meets all the criteria.
+          Checking that the number of strets (ny) is correct.", {
   world_test <- list(nx = 6, ny = "hola",
                      hor_walls = dplyr::tibble(x = 3, y = 1, lgth = 3),
                      ver_walls = dplyr::tibble(x = 3, y = 0, lgth = 1),
@@ -42,8 +48,8 @@ test_that("Checking provided nx and ny.", {
   expect_error(generar_mundo(world_test))
 })
 
-test_that("Checking provided karel_x, karel_y, karel_dir and karel_bag", {
-
+test_that("a new world provided as a list by the user meets all the criteria.
+          Checking that Karel's initial position (karel_x) is correct.", {
   # karel_x
   world_test <- list(nx = 6, ny = 4,
                      hor_walls = dplyr::tibble(x = 3, y = 1, lgth = 3),
@@ -68,7 +74,10 @@ test_that("Checking provided karel_x, karel_y, karel_dir and karel_bag", {
                      beepers_x = 2, beepers_y = 1, beepers_n = 1,
                      beepers_bag = Inf)
   expect_error(generar_mundo(world_test))
+})
 
+test_that("a new world provided as a list by the user meets all the criteria.
+          Checking that Karel's initial position (karel_y) is correct.", {
   # karel_y
   world_test <- list(nx = 6, ny = 4,
                      hor_walls = dplyr::tibble(x = 3, y = 1, lgth = 3),
@@ -93,8 +102,11 @@ test_that("Checking provided karel_x, karel_y, karel_dir and karel_bag", {
                      beepers_x = 2, beepers_y = 1, beepers_n = 1,
                      beepers_bag = Inf)
   expect_error(generar_mundo(world_test))
+})
 
-  # karel_dir
+test_that("a new world provided as a list by the user meets all the criteria.
+          Checking that Karel's initial direction (karel_dir) is correct.", {
+
   world_test <- list(nx = 6, ny = 4,
                      hor_walls = dplyr::tibble(x = 3, y = 1, lgth = 3),
                      ver_walls = dplyr::tibble(x = 3, y = 0, lgth = 1),
@@ -118,8 +130,12 @@ test_that("Checking provided karel_x, karel_y, karel_dir and karel_bag", {
[…]
