### 2024-02-14 9e40ea2
added badge to README as requiered by ropensci-review-bot
