### 2024-02-14 94acbf3
changed NEWS file to comply with rOpenSci suggestions
