+++ NEWS.md
@@ -1,24 +1,52 @@
-# karel 0.1.1.9000
+karel 0.1.1.9001 (2024-02-14)
+=============================
 
-* This the current development version of karel.
+### NEW FEATURES
 
-# karel 0.1.1
+* Unit tests were rewritten to make them more specific so that it's easier to detect the problem when a test fails. Large tests with lots of expectations were replaced by more but smaller ones.
+
+### MINOR FIXES
+
+* Fixed sentence in the *Getting started* vignettes.
+
+karel 0.1.1.9000 (2023-12-28)
+=============================
+
+### NEW FEATURES
+
+* The package was redesigned with a multilingual structure that allows new languages to be added.
+* Display of error messages was improved by using the cli package.
+
+### DOCUMENTATION FIXES
+
+* Documentation for several functions was improved.
+* The website was redesigned to be easy to access to Spanish and English materials.
+
+### OTHERS
+
+* Added continouos integration practices through GitHub Actions to build the website and to run package checks.
+
+karel 0.1.1
+===========
 
 * Fixed typo in DESCRIPTION and vignettes.
 * Updated vignettes names to indicate reading order and added descriptions.
 * Updated README with link to package website on CRAN and new badges.
 
-# karel 0.1.0
+karel 0.1.0
+===========
 
 * Version of first submission to CRAN.
 
-# karel 0.0.1.9000
+karel 0.0.1.9000
+================
 
 * Improved description file.
 * Added new worlds.
 * Added mode to prevent running more actions without generating a new world if user made a mistake. 
 * Added new vignettes with several examples and exercises.
 
-# karel 0.0.0.9000
+karel 0.0.0.9000
+================
 
 * This is the first development version of karel.
