### 2024-02-15 1c9cd1d
added interactivity tests with the vdiffr package and testthat snapshots
