### 2024-02-15 #620 @maurolepore
@mpru thanks a lot for this! It will make the reviewers's job easier.

> Upgraded the version number from 0.1.1.9000 to 0.1.1.9001 (not sure if this is ok).

I think it's OK. That's the kind of increment I see with `fledge::bum_version()` in my own packages.

> I didn't understand how to use Airtable to search for potential reviewers, could it be that the general public doesn't have access to the database and only editors?

I think you're right, and Airtable is only useful to editors with the right access permission.

I'll resume the search for reviewers.
