### 2024-03-15 #620 @vjimenez9
Aqui el resultado de mi revisión:

[PlatillaDeRevisionROpenSci.md](https://github.com/ropensci/software-review/files/14620152/PlatillaDeRevisionROpenSci.md)
