+++ vignettes/a_intro_progrbasics_en.Rmd
@@ -11,10 +11,7 @@ vignette: >
 ---
 
 ```{r, include = FALSE}
-knitr::opts_chunk$set(
-  collapse = TRUE,
-  comment = "#>"
-)
+knitr::opts_chunk$set(collapse = TRUE)
 ```
 
 # Some Ideas about Programming
@@ -25,9 +22,9 @@ All these devices with different types of computers have something in common: so
 
 Programming is a creative process: in many cases, a task can be accomplished by following different paths, and the programmer is the one who must imagine what those paths are and choose one. Some paths may be better than others, but in any case, the computer will simply follow the instructions devised by the programmer.
 
-Unfortunately, computers don't understand Spanish or any other human language. We have to give them instructions in a language they can understand. For that, we need to learn a *programming language*, which is simply an artificial language composed of a series of expressions that the computer can interpret. Computers interpret our instructions very literally, so when programming, we need to be very specific. We must follow the rules of the programming language and be clear in the instructions provided.
+Unfortunately, computers don't understand English or any other human language. We have to give them instructions in a language they can understand. For that, we need to learn a *programming language*, which is simply an artificial language composed of a series of expressions that the computer can interpret. Computers interpret our instructions very literally, so when programming, we need to be very specific. We must follow the rules of the programming language and be clear in the instructions provided.
 
-Now, why should we study programming if we are dedicated to Statistics? The activity of statistical professionals is entirely intertwined with the need to master computer tools that assist us in various stages of our work, from data collection and cleaning, to the application of different analysis methodologies, to the effective communication of results. That's why in this tutorial, we will study the basic concepts of this discipline, fostering the exercise of abstract and logical thinking necessary to effectively communicate with the computer and make it perform the tasks we need. The programming language used in this tutorial is R, widely used in the field of data science. However, we must emphasize that **this is not an R tutorial**, meaning we will not focus on learning the tools that this language provides for data analysis.
+Now, why should we learn to program when we want to analyze data, generate statistics, graphs and conclusions? The activity of statistical professionals, data science professionals, or anyone who requires data analysis is enhanced with the use of computer tools and programming capabilities that assist us in the different stages of our work, from the collection and purification of data sets, through the application of different analysis methodologies, to the effective communication of the results. That's why in this tutorial, we will study basic concepts of programming, fostering the exercise of abstract and logical thinking necessary to effectively communicate with the computer and make it perform the tasks we need. The programming language used in this tutorial is R, widely used in the field of data science. However, we must emphasize that **this is not an R tutorial**, meaning we will not focus on learning the tools that this language provides for data analysis.
 
 To incorporate some basic programming notions, we will make use of the help of **Karel**, a robot that lives and walks in rectangular **worlds**. We will ask Karel to perform certain activities, and for this, we have to program in R, so Karel will accompany us in the first steps we take with this language.
 
@@ -55,7 +52,9 @@ When faced with a problem, the first step is to devise an algorithm for its solu
 
 An algorithm presented in pseudocode could use terms taken from the Spanish language, which is a reasonable choice for sharing this strategy among people who communicate in this language. However, if we want to present our algorithm to someone who only speaks French, Spanish would no longer be a good choice, let alone if we want to present the algorithm to a computer. In order for a computer to understand our algorithm, we need to translate it into a *programming language*, which, as mentioned before, is an artificial language designed to express computations that can be carried out by electronic devices; in other words, it is a means of communication between humans and machines.
 
[…]
