+++ pkgdown.yml
@@ -1,4 +1,4 @@
-pandoc: 2.19.2
+pandoc: 3.1.11
 pkgdown: 2.0.7
 pkgdown_sha: ~
 articles:
@@ -14,7 +14,7 @@ articles:
   d_controlstructures_es: d_controlstructures_es.html
   e_examples_en: e_examples_en.html
   e_examples_es: e_examples_es.html
-last_built: 2024-02-15T06:41Z
+last_built: 2024-03-27T22:47Z
 urls:
   reference: https://mpru.github.io/karel/reference
   article: https://mpru.github.io/karel/articles
