+++ vignettes/e_examples_es.Rmd
@@ -13,7 +13,8 @@ vignette: >
 ```{r, include = FALSE}
 knitr::opts_chunk$set(
   collapse = TRUE,
-  comment = "#>", out.width = "60%", fig.align = "center"
+  comment = "",
+  prompt = TRUE, out.width = "60%", fig.align = "center"
 )
 ```
 
