+++ vignettes/aa_get_started_es.Rmd
@@ -12,7 +12,8 @@ vignette: >
 ```{r, include = FALSE}
 knitr::opts_chunk$set(
   collapse = TRUE,
-  comment = "#>",
+  comment = "",
+  prompt = TRUE,
   fig.align = "center"
 )
 ```
