+++ vignettes/e_examples_en.Rmd
@@ -13,7 +13,8 @@ vignette: >
 ```{r, include = FALSE}
 knitr::opts_chunk$set(
   collapse = TRUE,
-  comment = "#>"
+  comment = "",
+  prompt = TRUE, out.width = "60%", fig.align = "center"
 )
 ```
 
