+++ vignettes/b_meetingkarel_es.Rmd
@@ -13,7 +13,8 @@ vignette: >
 ```{r, include = FALSE}
 knitr::opts_chunk$set(
   collapse = TRUE,
-  comment = "#>",
+  comment = "",
+  prompt = TRUE,
   out.width='60%'
 )
 ```
