+++ vignettes/b_meetingkarel_en.Rmd
@@ -12,7 +12,8 @@ vignette: >
 ```{r, include = FALSE}
 knitr::opts_chunk$set(
   collapse = TRUE,
-  comment = "#>",
+  comment = "",
+  prompt = TRUE,
   out.width='60%'
 )
 ```
