### 2024-03-29 84dacc6
updated citation file to fix something noted by check_win
