+++ vignettes/a_intro_progrbasics_en.Rmd
@@ -68,7 +68,7 @@ We have defined an algorithm as a list of instructions to solve a problem. In th
 
 ## **R and RStudio**
 
-As mentioned earlier, we will put into practice programming concepts using **R**. Unlike many other languages, R is free and open-source software distributed under the *GNU GPLv2* license, which allows freedom to use it for any purpose, to see how it was made, how it works, and to modify it, distribute copies, and create contributions and additions for them to be available to others ^[**1**](https://chat.openai.com/#user-content-fn-2)^.
+As mentioned earlier, we will put into practice programming concepts using **R**. Unlike many other languages, R is free and open-source software distributed under the *GNU GPLv2* license, which allows freedom to use it for any purpose, to see how it was made, how it works, and to modify it, distribute copies, and create contributions and additions for them to be available to others.
 
 Although R will be our means of communication with the computer, we will use another program that provides some tools to facilitate our programming work. In other words, we will use an integrated development environment (IDE). An IDE is a program that makes coding easier by allowing us to manage multiple code files, visualize the working environment, use color highlighting for different parts of the code, employ autocomplete to write faster, explore help pages, implement debugging strategies, and even interactively execute instructions while viewing the results as we progress in analyzing or solving the problem. The most popularly used IDE for programming with R is RStudio, and it will be the program we'll be using all the time.
 
