### 2024-03-29 #620 @mpru
¡Hola @vjimenez9!

Muchísimas gracias por haber aceptado revisar este paquete y por el esfuerzo y dedicación puestos en esta tarea. Tus observaciones son muy útiles. A continuación voy a ir respondiéndolas:

**Documentación**

> Instrucciones de instalación
> Nota: En README no se especifica que no es posible instalarlo en Ubuntu, tampoco que se requiere tener instalado RUST

Acerca de la instalación en Ubuntu, no he podido reproducir este inconveniente. De hecho, el paquete ha sido desarrollado en Ubuntu, ya que es el sistema operativo con el que trabajo. Utilicé `rhub::check_on_ubuntu()` para evaluarlo en un medio externo y al parecer [la evaluación fue exitosa](https://builder.r-hub.io/status/original/karel_0.1.1.9001.tar.gz-824606cdd7fc41fa9e99892f1a11e99e).

De todas formas, veo en la salida que muestras que dice "Karel" con mayúscula, el paquete se llama "karel" con minúscula. ¿Puede que sea esa la razón por la cual no lo pudiste instalar?

Tampoco hemos detectado problemas de instalación en Windows o en Posit Cloud, que son los entornos más utilizamos en clase. Además de correr pkgcheck, he utilizado devtools para chequear el paquete con check_win, abajo pongo los enlaces, no parece haber habido algún problema relacionado a la instalación:

- [`devtools::check_win_devel()`]()
- [`devtools::check_win_oldrelease()`](https://win-builder.r-project.org/sw3av21OrRko)
- [`devtools::check_win_release()`](https://win-builder.r-project.org/2a9207FD36hS)

Con respecto a macOS, he corrido `devtools::check_mac_release()` y no parece haber detectado problemas con la instalación. Lamentablemente ni yo ni las personas en mi entorno usan macOS como para poder hacer pruebas locales, no sé cómo afrontar esta parte, agradezco sugerencias.

**Funcionalidad**

El punto *Directrices de empaque* no ha sido tildado, ¿podrías orientarme acerca de qué aspecto revisar? Entiendo que relacionado con esto debe estar el comentario de la revisón que dice:

> ¿El paquete cumple con la guía de embalaje rOpenSci? en general si

**Comentarios de la revisión**

> ¿El código cumple con los principios generales de la guía de revisión de Mozilla?
> No me queda claro si hubo alguna revisión por pares, o como dice la guia es un Flying Solo. Por otro lado me parece que la meta es clara y se cumple.

El paquete no ha tenido otra revisión por fuera de rOpenSci, pero participar del presente proceso de revisión de rOpenSci a través del programa Champions satisface este requisito.

> ¿Hay duplicación de código en el paquete que debería reducirse?
> No necesariamente se duplica, pero hay lugares donde se puede obtimizar. Por ejemplo donde se dibuja a Karel, se pueden definir un grupo de vectores que se asignen al tible en función de la dirección, en lugar de crear los 4 tibles por separado.
> Tambien cuando se hacen los chequeos ...se podria genear una función a la que se llama multiples veces con el objeto a checar...
> etc.

Cambiar la estructura de objetos diseñada para graficar a Karel en cada paso implicaría reformar todo el código base del paquete. En su uso no se me han presentado situaciones que me hicieran pensar que una reestructuración sea necesaria. ¿Consideras estrictamente necesario este cambio? ¿O es sólo una observación?

**Sección Mis notas**

> Averigundo un poco, encontré que No hay una versión para ubuntu. Sería util que lo dijeran desde el README.
> Luego hice la instalación en una macOS.
> Les dejo aqui los detalles de los inconvenientes que tuve para realizar la istalación (Solo con fines informativos):

Ver mi comentario anterior al respecto.

**Sugerencias de modificación en el texto**.

Con respecto a tus sugerencias de mejoras para el texto de los tutoriales, te agradezco por cada una de ellas y por haberlo leído con atención. Incorporé casi todas en el commit https://github.com/mpru/karel/commit/fb65a4c40118cff94a756968458f459ec31335a9, dejando de lado sólo algunas porque se refieren a cuestiones que en realidad estaban resueltas o presentes en otras partes, o por diferencia de criterio.

Bueno, esto es lo que tengo para responder con respecto a tu revisión. Te agradezco nuevamente por ĺa buena voluntad y disposición de evaluar mi trabajo, espero que podamos seguir trabajando hasta lograr el objetivo.
