### 2024-03-31 #620 @vjimenez9
Estimado @mpru:

Tienes razón,  es muy probable que el error fue por la K mayúscula de Karel en la instalación de ubuntu.  Esta fue la razón por la que no puse el tilde en "directrices de empaque". porque no me funcionaba en ubuntu y hubo complicaciones en mac, básicamente fue esa la razón.

Reestructurar el código para reducir duplicidad o simplificarlo, es una sugerencia,  y no un requisito obligatorio.  No tengo problema con que se quede como esta.  Finalmente funciona, y funciona bien.

Tu herramienta es muy bonita,  pero finalmente tengo ya varios años generando material para la enseñanza de programación y por eso me atreví a SUGERIR tan detalladamente en el texto, siempre con el objetivo de mejorarlo.  Me parece que la mayoría de lo vital quedo plasmado. Supongo que con esto terminan mis sugerencias y no me queda mas que felicitarte.
