### 2024-04-02 #620 @joelnitta
---

#### To-do list

- [ ] Add statement of need. The `Who is Karel?` section of the README hints at the need but does not describe it explicitly.

#### UI

Many of the error messages are preceded by `cli::cli_rule()`. While of course informative error messages are important, care should be taken that the package is not [overly chatty](https://ropensci.org/blog/2024/02/06/verbosity-control-packages/). I don't see why the `cli::cli_rule()` is needed.

For `expect_error()`, it is a good idea to indicate the expected error message using the `regexp` argument so that you know the error is actually the expected one and not another error. This also makes the intention of the test clearer, for example the first test of `test-set_up.R`.

There is a lot of repeated code in `test-set_up.R` where each list of world elements (`world_test`) is being defined slightly differently. I think this could be simplified by starting with a single `world_test_template` list, then modifying it for each test to check that the correct error message is generated. For example:

```
world_test_template <- list(
  nx = 6, ny = 4,
  hor_walls = dplyr::tibble(x = 3, y = 1, lgth = 3),
  ver_walls = dplyr::tibble(x = 3, y = 0, lgth = 1),
  karel_x = 1, karel_y = 1,
  karel_dir = 1,
  beepers_x = 2, beepers_y = 1, beepers_n = 1,
  beepers_bag = Inf)

test_that("The generic world template works correctly before modifying it
           in subsequent tests", {
  expect_no_error(generar_mundo(world_test_template))
  expect_no_error(generate_world(world_test_template))
})

test_that("Karel's initial direction is required when creating a new world
           provided as a list by the user.", {
  world_test <- world_test_template
  world_test$karel_x <- NULL
  expect_error(generate_world(world_test), "karel_x is missing")
})

test_that("The correct number of avenues (nx) is required when creating a newa
           world provided as a list by the user.", {
  world_test <- world_test_template
  world_test$nx <- 6.4
  expect_error(
    generate_world(world_test),
    "nx, the number of avenues, must be numeric of length 1")
})

```

It would be good to add a test coverage badge so we can verify what percentage of functions are included in tests.

#### Examples

I recommend use of `if (interactive())` instead of `if(FALSE)` for skipping example code like `run_actions()` that should not be run during checks. `if (interactive())` is more explicit in its intent than `if(FALSE)`, and will work when users copy-paste the example code. Sometimes CRAN checks can be picky about how you specify which code to skip in examples.

#### Multilingual documentation

As far as I know, this package is unique in its approach to multilingual functionality and documentation. I applaud the author for this commitment to breaking down linguistic barriers. However, there are several aspects to consider carefully here.

I am a little concerned about the multilingual aliases for function names. First, I think ultimately it could be counter-productive for the learners. For better or worse, English is the standard language used in programming. If the goal is to teach students programming, at some point they will have to deal with words that are in English. We can even see this in `karel` lessons: although the `karel` functions have names in Spanish, the learners still need to use base R code with syntax like `if()`, `else()`, `function()`, etc. It may be better to get them used to writing R syntax in English from the start instead of mixing English and Spanish (or another language). Furthermore, CRAN does not allow non-ASCII characters in R code, so it will be difficult or impossible to display function names in some non-English languages. Finally, using only one name per function makes the package easier to maintain. Currently, entire test files are duplicated and must be maintained due to having function aliases. However, as I mention below, one argument in favor of function aliases is that it allows you to maintain help files in a different languages... this is a complicated topic!

Apart from function names, I think localization of function UI and documentation is absolutely a plus and should be done. I see this as falling into three main categories: localization of UI, documentation of functions in the package (help files), and the package website. This package is pushing the limits of what R can do in terms of multilingual functionality, and as I describe below, mature solutions for each of these are not yet available in some cases.

For localizing the UI (function messages), I recommend the [potools](https://github.com/MichaelChirico/potools/) package (and PO files). This provides a very clean way to localize a package. Once it is set up, the user should have a seamless linguistic experience: if they are using a computer with a Spanish locale, all of the package messages will be in Spanish; same for English, etc. To get started with `potools`, I recommend this [blog post by Maëlle Salmon](https://masalmon.eu/2023/10/06/potools-mwe/). This is the localization aspect that has the most mature solution available.

@eliocamp is working on a package to localize help file contents, [rhelpi18n](https://github.com/eliocamp/rhelpi18n). It should eventually provide a similar seamless experience as UI localization with PO tools, but unfortunately does not seem ready for production use yet. So honestly I am not sure what to do here. I wish there was a way to provide a help file in an alternate language that did not require a different function name. Actually, this may be a point in favor of keeping function aliases: it actually allows you to provide help files in different languages.

For the website, it would be ideal if there was a "language button" that could be clicked to switch between languages on a given page. With the current setup, both languages (Spanish and English) are displayed in a single webpage, but a given user probably only needs to see one or the other. Also, while this approach works OK for two languages, it would likely become unusable if any more were added. It would be great if pkgdown fully supported multilingual websites, but unfortunately [this does not seem to be the case](https://github.com/r-lib/pkgdown/issues/2258). Although pkgdown can provide website elements in different languages [by setting a YAML parameter](https://pkgdown.r-lib.org/articles/pkgdown.html?q=language#language), it is limited to making a single webpage in one language. Since this is currently set to Spanish, various website elements appear in Spanish even though some of the content is in English. So, in absence of a genuine solution from pkgdown, I wonder if a work-around via forking the package and maintaining a website in a different language from there could be an option. This is not ideal, but it would scale better if multiple languages were to be implemented.

### Vignettes

This package is also unique in that much of the vignette contents are actually lessons teaching programming in R with `karel`, rather just than how to use the `karel` functions. I think it may be preferable to split out the lessons teaching R into their own website. This would make maintaining the package easier. I am not sure of the optimal format, but one that I can recommend is the [Carpentries lesson template (known as "the Workbench")](https://carpentries.github.io/sandpaper-docs/). I could easily see the vignette contents being converted into a Carpentries lesson and hosted on the [Carpentries Incubator](https://carpentries-incubator.org/). Please consider this.

I also noticed that much of the lesson content is focused on control flow. While this makes sense as a general feature of programming languages, R these days is often used for data analysis, which may not necessarily put as much emphasis on control flow. I personally almost never use `while()`, and try to use `apply` or `purrr::map_()` family functions instead of `for` loops. I wonder if there may be someway to use `karel` to help teach data analysis, as it is in relatively high demand.

### Other random comments

- When the gif plays, it is automatically looped. This makes it a bit hard to tell where the commands start and end, especially if Karel ends up in the same place she started. You may want to add "start" and "end" text to make this clearer. Another idea is to make the looping optional, since it is a bit confusing to the learner that Karel repeats her actions even though we have not explicitly used a `for()` loop etc.

- Why do the beepers have the number "1" on them?

- Under the [Available actions for Karel](https://mpru.github.io/karel/reference/index.html#available-actions-for-karel) section in English, there are two functions in Spanish, `poner_coso()` and `juntar_coso()`.

- The examples for the internal functions should use `karel:::`, not `karel::`. Because of this, the examples for `get_pkg_env()` and `plot_static_world()` fail.
