### 2024-04-03 #620 @maurolepore
Thanks @vjimenez9 and @joelnitta for your reviews!

@mpru I see you already responded to @vjimenez9's reivew (in https://github.com/ropensci/software-review/issues/620#issuecomment-2027821275). If that's your final response to that review, then please simply acknowledge it; else please add whatever you feel necessary and/or refer to https://github.com/ropensci/software-review/issues/620#issuecomment-2027821275 when appropriate.

Also please respond to @joelnitta's review.

We recommend responding within the next 2 weeks.

> In general aim for 3 weeks for review, 2 weeks for subsequent changes, and 1 week for reviewer approval of changes.
--https://devdevguide.netlify.app/softwarereview_editor#during-review
