### 2024-04-24 8496952
Rebuilt help pages and vignettes after typos were corrected
