+++ DESCRIPTION
@@ -30,7 +30,7 @@ LazyData: true
 Language:
   en,
   es
-RoxygenNote: 7.2.3
+RoxygenNote: 7.3.1
 Suggests: 
     testthat (>= 3.0.0),
     knitr,
