### 2024-04-24 9024f2c
Calls to internal functions in examples corrected
