### 2024-04-24 7bfce1e
Solved incorrect grouping of functions in documentation
