+++ man/actions.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/lang_wrappers_en.R, R/lang_wrappers_es.R
+% Please edit documentation in R/lang_wrappers_en.R
 \name{actions}
 \alias{actions}
 \alias{move}
@@ -8,8 +8,6 @@
 \alias{pick_beeper}
 \alias{turn_right}
 \alias{turn_around}
-\alias{poner_coso}
-\alias{juntar_coso}
 \title{Available actions for Karel}
 \usage{
 move()
@@ -23,10 +21,6 @@ pick_beeper()
 turn_right()
 
 turn_around()
-
-poner_coso()
-
-juntar_coso()
 }
 \value{
 These functions don't return anything, but make changes in Karel's
