### 2024-04-24 15272ee
Changed to FALSE the default value for the loop argument in run_actions
