+++ R/lang_wrappers_en.R
@@ -78,7 +78,7 @@ generate_world <- function(world) .generate_world(world, lang = "en")
 #' Karel since its world was generated by \code{generate_world}.
 #'
 #' @param loop A logical value TRUE or FALSE indicating if the animation should
-#' repeat itself after finished or not (defaults to TRUE).
+#' repeat itself after finished or not (defaults to FALSE).
 #'
 #' @return Produces an animation with \code{gganimate}.
 #'
@@ -94,7 +94,7 @@ generate_world <- function(world) .generate_world(world, lang = "en")
 #'
 #' @export
 #'
-run_actions <- function(loop = TRUE) .run_actions(loop = loop, lang = "en")
+run_actions <- function(loop = FALSE) .run_actions(loop = loop, lang = "en")
 
 # ------------------------------------------------------------------------------
 
