### 2024-04-25 cb143b5
Changed if (FALSE) to if (interactive()) in examples
