+++ man/acciones.Rd
@@ -40,7 +40,7 @@ avanzar()
 juntar_coso()
 girar_izquierda()
 poner_coso()
-if (FALSE) ejecutar_acciones()
+if (interactive()) ejecutar_acciones()
 
 }
 \seealso{
