+++ man/actions.Rd
@@ -39,7 +39,7 @@ move()
 pick_beeper()
 turn_left()
 put_beeper()
-if (FALSE) run_actions()
+if (interactive()) run_actions()
 
 }
 \seealso{
