### 2024-04-25 a644c01
Removed cli::cli_rule() in all error messages
