+++ R/helpers.R
@@ -99,14 +99,12 @@
 .plot_static_world <- function(time, lang) {
 
   if (time < 1) {
-    cli::cli_rule()
     cli::cli_abort(call = NULL, message = c(
       "x" = message_texts[[lang]]$time_less_than_one_x,
       ">" = message_texts[[lang]]$time_less_than_one_arrow)
     )
   }
   if (time > pkg_env$moment) {
-    cli::cli_rule()
     cli::cli_abort(call = NULL, message = c(
       "x" = message_texts[[lang]]$time_too_big_x,
       ">" = paste0(message_texts[[lang]]$time_too_big_arrow, " ",
@@ -294,7 +292,6 @@ check_user_world <- function(world, lang) {
                 "beepers_bag")
   for (elem in elements) {
     if (!elem %in% names(world)) {
-      cli::cli_rule()
       cli::cli_abort(call = NULL, message = c(
         "!" = message_texts[[lang]]$check_user_world_general,
         "x" = paste(elem,
@@ -309,7 +306,6 @@ check_user_world <- function(world, lang) {
   # if the length is greater than 1 and that gives a warning, so I first check
   # the length
   if (length(world$nx) != 1 | !is.numeric(world$nx)) {
-    cli::cli_rule()
     cli::cli_abort(call = NULL, message = c(
       "!" = message_texts[[lang]]$check_user_world_general,
       "x" = message_texts[[lang]]$check_user_world_wrong_nx,
@@ -317,7 +313,6 @@ check_user_world <- function(world, lang) {
     )
   }
   if (world$nx %% 1 != 0) {
-    cli::cli_rule()
     cli::cli_abort(call = NULL, message = c(
       "!" = message_texts[[lang]]$check_user_world_general,
       "x" = message_texts[[lang]]$check_user_world_wrong_nx,
@@ -327,7 +322,6 @@ check_user_world <- function(world, lang) {
 
   # Check ny
   if (length(world$ny) != 1 | !is.numeric(world$ny)) {
-    cli::cli_rule()
     cli::cli_abort(call = NULL, message = c(
       "!" = message_texts[[lang]]$check_user_world_general,
       "x" = message_texts[[lang]]$check_user_world_wrong_ny,
@@ -335,7 +329,6 @@ check_user_world <- function(world, lang) {
     )
   }
   if (world$ny %% 1 != 0) {
-    cli::cli_rule()
     cli::cli_abort(call = NULL, message = c(
       "!" = message_texts[[lang]]$check_user_world_general,
       "x" = message_texts[[lang]]$check_user_world_wrong_ny,
@@ -345,7 +338,6 @@ check_user_world <- function(world, lang) {
 
   # Check karel_x
   if (length(world$karel_x) != 1 | !is.numeric(world$karel_x)) {
-    cli::cli_rule()
     cli::cli_abort(call = NULL, message = c(
       "!" = message_texts[[lang]]$check_user_world_general,
       "x" = message_texts[[lang]]$check_user_world_wrong_karel_x,
@@ -353,7 +345,6 @@ check_user_world <- function(world, lang) {
     )
   }
   if (world$karel_x %% 1 != 0) {
-    cli::cli_rule()
     cli::cli_abort(call = NULL, message = c(
       "!" = message_texts[[lang]]$check_user_world_general,
       "x" = message_texts[[lang]]$check_user_world_wrong_karel_x,
@@ -361,7 +352,6 @@ check_user_world <- function(world, lang) {
     )
   }
   if (world$karel_x < 1 | world$karel_x > world$nx) {
-    cli::cli_rule()
     cli::cli_abort(call = NULL, message = c(
       "!" = message_texts[[lang]]$check_user_world_general,
[…]
