+++ tests/testthat/test-set_up.R
@@ -1,352 +1,303 @@
-test_that("a new world provided as a list by the user meets all the criteria.
-          Checking that Karel's initial direction is indicated.", {
-  world_test <- list(nx = 6, ny = 4,
-                     hor_walls = dplyr::tibble(x = 3, y = 1, lgth = 3),
-                     ver_walls = dplyr::tibble(x = 3, y = 0, lgth = 1),
-                     karel_x = 1, karel_y = 1,
-                     # karel_dir = 1,
-                     beepers_x = 2, beepers_y = 1, beepers_n = 1,
-                     beepers_bag = Inf)
-  expect_error(generar_mundo(world_test))
+# generic world template to be used in tests
+world_test_template <- list(
+  nx = 6, ny = 4,
+  hor_walls = dplyr::tibble(x = 3, y = 1, lgth = 3),
+  ver_walls = dplyr::tibble(x = 3, y = 0, lgth = 1),
+  karel_x = 1, karel_y = 1,
+  karel_dir = 1,
+  beepers_x = 2, beepers_y = 1, beepers_n = 1,
+  beepers_bag = Inf
+)
+
+test_that("The generic world template works correctly before modifying it
+          in subsequent tests", {
+ expect_no_error(generar_mundo(world_test_template))
+ expect_no_error(generate_world(world_test_template))
 })
 
-test_that("a new world provided as a list by the user meets all the criteria.
-          Checking that the number of avenues (nx) is correct.", {
-  world_test <- list(nx = 6.4, ny = 4,
-                     hor_walls = dplyr::tibble(x = 3, y = 1, lgth = 3),
-                     ver_walls = dplyr::tibble(x = 3, y = 0, lgth = 1),
-                     karel_x = 1, karel_y = 1, karel_dir = 1,
-                     beepers_x = 2, beepers_y = 1, beepers_n = 1,
-                     beepers_bag = Inf)
-  expect_error(generar_mundo(world_test))
+test_that("Karel's initial direction is required when creating a new world
+           provided as a list by the user.", {
+   world_test <- world_test_template
+   world_test$karel_dir <- NULL
+   expect_error(generate_world(world_test), "karel_dir is missing")
 
-  world_test <- list(nx = 6:8, ny = 4,
-                     hor_walls = dplyr::tibble(x = 3, y = 1, lgth = 3),
-                     ver_walls = dplyr::tibble(x = 3, y = 0, lgth = 1),
-                     karel_x = 1, karel_y = 1, karel_dir = 1,
-                     beepers_x = 2, beepers_y = 1, beepers_n = 1,
-                     beepers_bag = Inf)
-  expect_error(generar_mundo(world_test))
+   world_test$karel_dir <- 1:4
+   expect_error(generate_world(world_test),
+      "karel_dir, Karel's initial direction, must be numeric of length 1")
+
+   world_test$karel_dir <- 5
+   expect_error(generate_world(world_test), "either 1, 2, 3 or 4")
+
+   world_test$karel_dir <- 1.5
+   expect_error(generate_world(world_test),
+            "karel_dir, Karel's initial direction, must be numeric of length 1")
 })
 
-test_that("a new world provided as a list by the user meets all the criteria.
-          Checking that the number of strets (ny) is correct.", {
-  world_test <- list(nx = 6, ny = "hola",
-                     hor_walls = dplyr::tibble(x = 3, y = 1, lgth = 3),
-                     ver_walls = dplyr::tibble(x = 3, y = 0, lgth = 1),
-                     karel_x = 1, karel_y = 1, karel_dir = 1,
-                     beepers_x = 2, beepers_y = 1, beepers_n = 1,
-                     beepers_bag = Inf)
-  expect_error(generar_mundo(world_test))
[…]
