### 2024-05-09 9c1b69c
Added coverage badge in readme and github action to monitor it
