+++ README.Rmd
@@ -22,7 +22,7 @@ knitr::opts_chunk$set(
 [![pkgcheck](https://github.com/mpru/karel/workflows/pkgcheck/badge.svg)](https://github.com/mpru/karel/actions?query=workflow%3Apkgcheck)
 [![R-CMD-check](https://github.com/mpru/karel/workflows/R-CMD-check/badge.svg)](https://github.com/mpru/karel/actions)
 [![Status at rOpenSci Software Peer Review](https://badges.ropensci.org/620_status.svg)](https://github.com/ropensci/software-review/issues/620)
-
+[![Codecov test coverage](https://codecov.io/gh/mpru/karel/branch/master/graph/badge.svg)](https://app.codecov.io/gh/mpru/karel?branch=master)
 <!-- badges: end -->
 
 ## ¿Quién es Karel? / Who is Karel?
