+++ README.md
@@ -16,7 +16,8 @@ developed.](https://www.repostatus.org/badges/latest/active.svg)](https://www.re
 [![R-CMD-check](https://github.com/mpru/karel/workflows/R-CMD-check/badge.svg)](https://github.com/mpru/karel/actions)
 [![Status at rOpenSci Software Peer
 Review](https://badges.ropensci.org/620_status.svg)](https://github.com/ropensci/software-review/issues/620)
-
+[![Codecov test
+coverage](https://codecov.io/gh/mpru/karel/branch/master/graph/badge.svg)](https://app.codecov.io/gh/mpru/karel?branch=master)
 <!-- badges: end -->
 
 ## ¿Quién es Karel? / Who is Karel?
