### 2024-05-09 #620 @mpru
@maurolepore, apologies for the delay, here I acknowledge my previous response to Veronica and proceed to respond to Joel.

## Response to @vjimenez9

My response to @vjimenez9 was stated in [this comment](https://github.com/ropensci/software-review/issues/620#issuecomment-2027821275). I really thank you again for your comments and suggestions.

## Response to @joelnitta

Thank you very much for all the work and commitment put into reviewing my package. I really appreciate all the comments and corrections made, many of which I would have liked to have read during the initial stages of the package's development, since they refer to structural issues of the package. Below I am going to respond to your comments one by one, indicating the changes applied to the package or the reason why I have not implemented, at least for now, some of the suggestions.

### Review Comments
#### To-do list
* [ ]  Add statement of need. The `Who is Karel?` section of the README hints at the need but does not describe it explicitly.

**Response**. _Updated README in https://github.com/mpru/karel/commit/d07dfcbaebdb15c85068264e7fda4c534a2c456a to try to address it more explicitly._

#### UI
Many of the error messages are preceded by `cli::cli_rule()`. While of course informative error messages are important, care should be taken that the package is not [overly chatty](https://ropensci.org/blog/2024/02/06/verbosity-control-packages/). I don't see why the `cli::cli_rule()` is needed.

**Response**. _Removed all calls to `cli_rule()` in https://github.com/mpru/karel/commit/a644c01f24e72202e132580d30c614e11ef417fa_

#### Tests
For `expect_error()`, it is a good idea to indicate the expected error message using the `regexp` argument so that you know the error is actually the expected one and not another error. This also makes the intention of the test clearer, for example the first test of `test-set_up.R`.

There is a lot of repeated code in `test-set_up.R` where each list of world elements (`world_test`) is being defined slightly differently. I think this could be simplified by starting with a single `world_test_template` list, then modifying it for each test to check that the correct error message is generated. For example:

```
world_test_template <- list(
  nx = 6, ny = 4,
  hor_walls = dplyr::tibble(x = 3, y = 1, lgth = 3),
  ver_walls = dplyr::tibble(x = 3, y = 0, lgth = 1),
  karel_x = 1, karel_y = 1,
  karel_dir = 1,
  beepers_x = 2, beepers_y = 1, beepers_n = 1,
  beepers_bag = Inf)

test_that("The generic world template works correctly before modifying it
           in subsequent tests", {
  expect_no_error(generar_mundo(world_test_template))
  expect_no_error(generate_world(world_test_template))
})

test_that("Karel's initial direction is required when creating a new world
           provided as a list by the user.", {
  world_test <- world_test_template
  world_test$karel_x <- NULL
  expect_error(generate_world(world_test), "karel_x is missing")
})

test_that("The correct number of avenues (nx) is required when creating a newa
           world provided as a list by the user.", {
  world_test <- world_test_template
  world_test$nx <- 6.4
  expect_error(
    generate_world(world_test),
    "nx, the number of avenues, must be numeric of length 1")
})
```

**Response**: _Changed tests to implement this in https://github.com/mpru/karel/commit/ae770bf786724da4ba9b73393a45e3ca2f9218fe and https://github.com/mpru/karel/commit/a995fa86dce5687c191764316031e592bd58efd2_

It would be good to add a test coverage badge so we can verify what percentage of functions are included in tests.

**Response**: _Implemented in https://github.com/mpru/karel/commit/9c1b69c4a6888a53c6db2f1668e861ffd31c6267_

#### Examples
I recommend use of `if (interactive())` instead of `if(FALSE)` for skipping example code like `run_actions()` that should not be run during checks. `if (interactive())` is more explicit in its intent than `if(FALSE)`, and will work when users copy-paste the example code. Sometimes CRAN checks can be picky about how you specify which code to skip in examples.

**Response**. _Fixed in https://github.com/mpru/karel/commit/cb143b56abe312f92968eba8b7d10bdbc4ceae3c_

#### Multilingual documentation
As far as I know, this package is unique in its approach to multilingual functionality and documentation. I applaud the author for this commitment to breaking down linguistic barriers. However, there are several aspects to consider carefully here.

I am a little concerned about the multilingual aliases for function names. First, I think ultimately it could be counter-productive for the learners. For better or worse, English is the standard language used in programming. If the goal is to teach students programming, at some point they will have to deal with words that are in English. We can even see this in `karel` lessons: although the `karel` functions have names in Spanish, the learners still need to use base R code with syntax like `if()`, `else()`, `function()`, etc. It may be better to get them used to writing R syntax in English from the start instead of mixing English and Spanish (or another language). Furthermore, CRAN does not allow non-ASCII characters in R code, so it will be difficult or impossible to display function names in some non-English languages. Finally, using only one name per function makes the package easier to maintain. Currently, entire test files are duplicated and must be maintained due to having function aliases. However, as I mention below, one argument in favor of function aliases is that it allows you to maintain help files in a different languages... this is a complicated topic!

**Response**. _I share your point of view. For better or worse, and sooner or later, our students must learn to use English, not only for programming, but because much of the new literature in the discipline is written in English. However, for the context in which we use this package (or in which I think others may use it, I write more about this later), I think there is an important difference between using and learning just a few English terms (if, else, function, etc.) and having everything else in Spanish, and having absolutely everything in English. With aliases, students can think about and discuss what actions Karel should take to solve a problem and write them down in their own language (e.g., girar a la derecha, juntar, poner), while also being able to consult help and see examples in their own language. I understand that this does not solve the need for English in the medium term, but I like having the possibility of using the environment mostly in Spanish as a first approach._

Apart from function names, I think localization of function UI and documentation is absolutely a plus and should be done. I see this as falling into three main categories: localization of UI, documentation of functions in the package (help files), and the package website. This package is pushing the limits of what R can do in terms of multilingual functionality, and as I describe below, mature solutions for each of these are not yet available in some cases.

For localizing the UI (function messages), I recommend the [potools](https://github.com/MichaelChirico/potools/) package (and PO files). This provides a very clean way to localize a package. Once it is set up, the user should have a seamless linguistic experience: if they are using a computer with a Spanish locale, all of the package messages will be in Spanish; same for English, etc. To get started with `potools`, I recommend this [blog post by Maëlle Salmon](https://masalmon.eu/2023/10/06/potools-mwe/). This is the localization aspect that has the most mature solution available.

**Response**: _Thanks for commenting on the `potools` package, I didn't know about it. I think it's something I could explore, and I'd like to, although this would involve changing the current structure of the package extensively and in the short term I don't think I can do it. However, this approach goes hand in hand with having all function names written in one language (which should be English, following the previous reasoning that the rest of the statements like if, function, for, etc., are in English), but I really want and like the names of the functions to be in Spanish, and to use them in Spanish with my students, even though this incurs the alias system with the limitations that we discussed before._

@eliocamp is working on a package to localize help file contents, [rhelpi18n](https://github.com/eliocamp/rhelpi18n). It should eventually provide a similar seamless experience as UI localization with PO tools, but unfortunately does not seem ready for production use yet. So honestly I am not sure what to do here. I wish there was a way to provide a help file in an alternate language that did not require a different function name. Actually, this may be a point in favor of keeping function aliases: it actually allows you to provide help files in different languages.

**Response**: _What a nice project you mentioned. I seem to understand that the system proposed by the `rhelpi18n` package requires writing the documentation in different languages, which is also what I do now in my package, but saving these translations in different packages. That has the disadvantage of having to install one more package depending on the language you want to use it in, but it is an advantage because it is easier to maintain, and because eventually including more and more languages under the current karel concept would make the package become heavier than allowed. If this approach matures we could use it to have translations of the help pages, but again we would still have a single English name for each function, which I consider to be a limitation. I think I prefer karel's current approach for the moment._

For the website, it would be ideal if there was a "language button" that could be clicked to switch between languages on a given page. With the current setup, both languages (Spanish and English) are displayed in a single webpage, but a given user probably only needs to see one or the other. Also, while this approach works OK for two languages, it would likely become unusable if any more were added. It would be great if pkgdown fully supported multilingual websites, but unfortunately [this does not seem to be the case](https://github.com/r-lib/pkgdown/issues/2258). Although pkgdown can provide website elements in different languages [by setting a YAML parameter](https://pkgdown.r-lib.org/articles/pkgdown.html?q=language#language), it is limited to making a single webpage in one language. Since this is currently set to Spanish, various website elements appear in Spanish even though some of the content is in English. So, in absence of a genuine solution from pkgdown, I wonder if a work-around via forking the package and maintaining a website in a different language from there could be an option. This is not ideal, but it would scale better if multiple languages were to be implemented.

**Response**. _Yes, every time I see the website I think exactly that, with two languages it is fairly ok, but if someone really wanted to keep adding languages, it would quickly become impractical. Hopefully at some point pkgdown will support support for multilingual pages, that would be great. It seems to me that the forking idea could work if more languages are added, including links between the different webpages, although I'm not sure about the source code living in several kind of "official" repos. I think for now I'm going to keep the site as planned now._

#### Vignettes
This package is also unique in that much of the vignette contents are actually lessons teaching programming in R with `karel`, rather just than how to use the `karel` functions. I think it may be preferable to split out the lessons teaching R into their own website. This would make maintaining the package easier. I am not sure of the optimal format, but one that I can recommend is the [Carpentries lesson template (known as "the Workbench")](https://carpentries.github.io/sandpaper-docs/). I could easily see the vignette contents being converted into a Carpentries lesson and hosted on the [Carpentries Incubator](https://carpentries-incubator.org/). Please consider this.

**Response:** _I agree with this observation, I made use of the vignettes in a way that is outside the norm for R packages, turning them into lessons. In fact, I use these vignettes as study material for the students I use this material with. This also turns out to be an extra burden to achieve the objective of having all the documentation multilingual, adding a new language would mean translating all the vignettes. However, for the moment I prioritize making this project self-contained, and having my students rely on a single web page for the package and their lessons. I'd be interested in exploring your suggestions about including this in the Carpentries Incubator in the medium term, but it's not something I can pursue right away._

I also noticed that much of the lesson content is focused on control flow. While this makes sense as a general feature of programming languages, R these days is often used for data analysis, which may not necessarily put as much emphasis on control flow. I personally almost never use `while()`, and try to use `apply` or `purrr::map_()` family functions instead of `for` loops. I wonder if there may be someway to use `karel` to help teach data analysis, as it is in relatively high demand.

**Response:** _What you notice has been chosen on purpose. To better answer, I must give some context about my use of this package. At my University, this package has been used in an informal pre-university course for some years now for students who will enter the Bachelor's degree in Statistics and who do not have any prior knowledge of programming. The objective is to teach them general programming concepts (applicable to any language) but in the main language that they will begin to use in the immediate future in their studies, R. That's why tools for data analysis or to improve some aspects of coding (like the apply or map examples you mention) are not included in the vignettes. They study all this and much more in the degree. In this course we only see these general programming ideas, not oriented to data analysis, in the environment that they will later use for data analysis. Perhaps the question that I should asked to myself is whether, thinking that the vignettes/lessons can be used in other contexts, it should not include other notions, but I think not, since in this way it remains limited and faithful to its original purpose._

#### Other random comments
* When the gif plays, it is automatically looped. This makes it a bit hard to tell where the commands start and end, especially if Karel ends up in the same place she started. You may want to add "start" and "end" text to make this clearer. Another idea is to make the looping optional, since it is a bit confusing to the learner that Karel repeats her actions even though we have not explicitly used a `for()` loop etc.

	**Response:** _the looping was already optional, as I faced this problem before. I couldn't make my mind about this, and ended up setting to `TRUE` the default value for the `repeat` argument in the `.run_actions` internal function and its wrapers. Now that you mention this, I changed it to `FALSE` in https://github.com/mpru/karel/commit/15272eec9cbdd4ae7f91aeec30331c7a65ef2477_

* Why do the beepers have the number "1" on them?

	**Response:** _Because in each position there can be more than one beeper and this number indicates the quantity present there (if there's no beeper, then it's not ploted, otherwise the number tells how many there are in that spot). In some examples of the tutorials there are very interesting problems and exercises to solve taking this characteristic into account, such as detecting how many beepers there are in a position and duplicating them._

* Under the [Available actions for Karel](https://mpru.github.io/karel/reference/index.html#available-actions-for-karel) section in English, there are two functions in Spanish, `poner_coso()` and `juntar_coso()`.

	**Response:** _solved in https://github.com/mpru/karel/commit/7bfce1ee3864eaefc3af698691f8b33f8820de16_

* The examples for the internal functions should use `karel:::`, not `karel::`. Because of this, the examples for `get_pkg_env()` and `plot_static_world()` fail.

	**Response:** _solved in https://github.com/mpru/karel/commit/9024f2c114f530abe6439bd3f2ea113a1c8f968a_

* Please note I submitted a [PR](https://github.com/mpru/karel/pull/16/) correcting some typos

	**Response**: _I accepted it, thank you for having done such a detailed reading and having detected these typos._
