### 2024-05-10 #620 @maurolepore
¡Gracias @mpru!

@vjimenez9 y @joelnitta, ¿podrían por favor informarnos si recomiendan más cambios o indicar su aprobación utilizando esta  [plantilla](https://devdevguide.netlify.app/approval2template)?

--

Thanks @mpru!

@vjimenez9 and @joelnitta can you please let us know if you recommend further changes or indicate your approval using this [template](https://devdevguide.netlify.app/approval2template)?
