### 2024-06-19 #620 @maurolepore
@vjimenez9

Revisando los comentarios veo este comentario por el que asumo tu aprovacion:

> Me parece que la mayoría de lo vital quedo plasmado. Supongo que con esto terminan mis sugerencias y no me queda mas que felicitarte. -- @vjimenez9 en https://github.com/ropensci/software-review/issues/620#issuecomment-2028888781

Para no demorar la publicacion voy a formalizar la aprobacion. Si tenes alguna sugerencia mas, es bienvenida pero sugiero que @mpru la considere luego de la publicacion.

De todos modos, cuando puedas, por favor completa esta [plantilla](https://devdevguide.netlify.app/approval2template) asi registramos formalmente tu esfuerzo.

Muchisimas gracias  por tu trabajo.

----

@vjimenez9

Reviewing the comments, I see this one from which I assume your approval (translated with ChatGTP):

> It seems to me that most of the essential points have been addressed. I suppose this concludes my suggestions, and I can only congratulate you. -- @vjimenez9 on https://github.com/ropensci/software-review/issues/620#issuecomment-2028888781

To avoid delaying the publication, I will formalize the approval. If you have any further suggestions, they are welcome but I suggest that @mpru consider them after the publication.

In any case, when you can, please complete this [template](https://devdevguide.netlify.app/approval2template) so we can formally record your efforts.

Thank you very much for your work.
