### 2024-06-19 #620 @ropensci-review-bot
Approved! Thanks @mpru for submitting and @vjimenez9, @joelnitta for your reviews! :grin:
