### 2024-07-23 #620 @maurolepore
@mpru I see the package is approved but [it seems it hasn't been moved to the ropensci](https://github.com/orgs/ropensci/repositories?q=karel) organization, right?

Can we support in any way?

cc @yabellini
