### 2024-07-27 #620 @mpru
Hola @maurolepore, gracias por tus mensajes y perdón por mi ausencia. La aprobación del coincidió justo con el nacimiento de mi hija y tuve que dejar en pausa algunas cosas.

Ayer comencé a completar los pasos mencionados arriba y en el proceso me surgió una duda, en la que tal vez @yabellini  pueda ayudar. Este paquete no entraba en ninguna de las categorías de alcances de rOpenSci pero fue aceptado en el marco del programa Champions para pasar por el proceso de revisión, tal como hemos hecho. Me pregunto ahora si después de la revisión y aprobación, igualmente debemos transferirlo a rOpenSci o no por esta particularidad. Quería recordar este aspecto antes de continuar.

Estoy en la parte referida a la nueva publicación del pkgdown website, espero confirmación para continuar.

¡Gracias!
