### 2024-08-14 #620 @yabellini
Hola, volvi de mis vacaciones y me estoy poniendo al dia. Disculpen la tardanza en volver al tema.

La idea original era poder enviarlo a JOSE despues de nuestra revision, pero eso implicaria escribir un paper (corto) pero aun implica escribirlo y debe ser en Ingles.  JOSE aplicaria un fast track para este paper porque ya fue revisado por nosotros y como yo soy editora alli podria encargarme del proceso.

Independiente del punto anterior, voy a volver a preguntar sobre el traspaso a la organziacion de rOpenSci, lo que si no creo que podamos hacer es crear la documentación con el workflow de rOpenSci, especialmente por la naturaleza multilingue de la documentacion.
