### 2024-08-22 #620 @mpru
Gracias, Yani, voy a ir trabajando en el borrador del paper para JOSE y mientras tanto sigo dejando en pausa lo de la transferencia del repo, que por ahora quedó en ropensci, cualquier cosa lo volvemos.
