### 2024-08-22 #620 @yabellini
Hola! Ya lo hablamos con el equipo y esta bien que el paquete este en la organizacion de rOpenSci.  Aqui ya esta la documentacion generada: https://docs.ropensci.org/karel/

Mas adelante estaremos resolviendo como agregarle los badges a los paquetes del programa de campeones para indicar que fueron parte del programa y con eso queda claro porque es parte de nuestra organizacion aun estando fuera de nuestro alcance.

Con respecto del paper avísame cualquier duda que tengas al respecto.

Felicitaciones @mpru por el trabajo realizado!
